+++ test/test_group_tableaux.jl
@@ -1,6 +1,5 @@
 @testitem "group theory tools" begin
     using Test
-
     using Random
     using QuantumClifford