+++ CHANGELOG.md
@@ -5,6 +5,16 @@
 
 # News
 
+## v0.9.12 - 2024-10-18
+
+- Minor compat fixes for julia 1.11 in the handling of `hgp`
+
+## v0.9.11 - 2024-09-27
+
+- `hcat` of Tableaux objects
+- `QuantumReedMuller` codes added to the ECC module
+- **(breaking)** change the convention for how to provide a representation function in the constructor of `LPCode` -- strictly speaking a breaking change, but this is not an API that is publicly used in practice
+
 ## v0.9.10 - 2024-09-26
 
 - The lifted product class of quantum LDPC codes is implemented in the ECC submodule.